#6 Feature - Add Merchants List and Search UI
1. Change file structure
2. Add Merchants List adn Search
