--- src/views/Merchants/MerchantForm.js
@@ -1,8 +1,0 @@
-import React from "react";
-const MerchantForm = () => {
-  return (
-    <div>Merchant Form</div>
-  );
-};
-
-export default MerchantForm;
