--- src/views/Merchants/MerchantPage.js
@@ -1,8 +1,0 @@
-import React from "react";
-const MerchantPage = () => {
-  return (
-    <div>Merchant Page</div>
-  );
-};
-
-export default MerchantPage;
