+++ src/views/Merchants/MerchantsForm.js
@@ -1,0 +1,8 @@
+import React from "react";
+const AccountForm = () => {
+  return (
+    <div>Account Form</div>
+  );
+};
+
+export default AccountForm;
